Refactored TripPatternMapper. Seperated out StopTimesMapper
