+++ src/test/java/org/opentripplanner/gtfs/mapping/StopTimesMapperTest.java
@@ -15,7 +15,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-public class StopTimeMapperTest {
+public class StopTimesMapperTest {
     private static final AgencyAndId AGENCY_AND_ID = new AgencyAndId("A", "1");
 
     private static final Integer ID = 45;
